refactor(promise-kit): move src/promiseKit to index
