+++ packages/promise-kit/index.js
